pyproject.toml
[project]
name = "apex-auto-app"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.0",
    "fastmcp>=2.12.4",
    "greenlet>=3.2.4",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.37.0",
]
